pyproject.toml
[tool.poetry]
name = "random-music"
version = "0.1.0"
description = "Fill up a directory (e.g.: a USB key) with random songs from a source directory."
authors = ["Michel Lavoie <lavoie.michel@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
rich-click = "^1.8.3"
mutagen = "^1.47.0"
PyYAML = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
